client/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pika==1.3.2
python-dotenv==1.0.0
cos-python-sdk-v5==1.9.28
Pillow==10.2.0
